class for song lists + filter for WhatsApp audios
