+++ app/src/main/java/com/hpixel/dreamusicplayer/MainActivity.kt
@@ -3,8 +3,7 @@ package com.hpixel.dreamusicplayer
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.widget.ListView
-import com.hpixel.dreamusicplayer.model.Files
-import com.hpixel.dreamusicplayer.model.Song
+import com.hpixel.dreamusicplayer.controller.SongListProvider
 import com.hpixel.dreamusicplayer.controller.SongsArrayAdaptor
 
 class MainActivity : AppCompatActivity() {
@@ -14,35 +13,14 @@ class MainActivity : AppCompatActivity() {
 		setContentView(R.layout.activity_main)
 
         val context = this.applicationContext
-        val songs = songList()
+
+        //filters out
+        val listProvider = SongListProvider()
+        val listFilter = SongListProvider.WHATSAPP_AUDIO_FILTER
+        val songList = listProvider.songList(context, listFilter)
+
         val songListView = findViewById(R.id.main_songList) as ListView
-        val songListArrayAdapter = SongsArrayAdaptor(context, songs)
+        val songListArrayAdapter = SongsArrayAdaptor(context, songList)
         songListView.adapter = songListArrayAdapter
 	}
-
-    fun songList(): ArrayList<Song> {
-        val files = Files()
-
-        val context = this.applicationContext
-        val separator = "||"
-        val songQuery = Song.query
-
-        val songArray = files.getSongs(context, separator, songQuery)
-
-        val songList = arrayListOf<Song>()
-
-        for (song in songArray){
-            val newSong = Song(song, separator)
-            songList.add(newSong)
-        }
-
-        //sort "alphabetically" by unicode value
-        songList.sortBy { song -> song.title }
-
-        //due to a possible "floating action button" that may be added later
-        val anEmptySpaceInTheEnd = Song(0, "", "", "", "", "", 0)
-        songList.add(anEmptySpaceInTheEnd)
-
-        return songList
-    }
 }
